[IndexedDB]: Explicit commit error timing handling.
This cl will handle the issue of when to abort an explicitly commited
transaction with errored requests and when to continue with the commit.
The rough approach will be to keep a tally of errors sent from the
backend to the front end and a tally of errors handled by the front end
and compare the two at commit time. If they do not match, the
transaction should be aborted.

Bug: 911877
Change-Id: I7ea7b9e20c70528de3f363e961f87a3d8f5798d3
GitHub Discussion: https://github.com/w3c/IndexedDB/pull/242