Move views array from XRFrame to XRViewerPose.
This is to match the latest version of the web XR spec for XRViewerPose.

The getViewerPose method makes a copy of the session's views relative to the reference space it was
queried with and stores it in the returned XRViewerPose object.

Updated the layout tests that need to access the views array.
Some of those tests required further adjustment to get an XRReferenceSpace object to pass to getViewerPose.

Bug: 915021
Change-Id: Ib0a8203034780b3936cf2e2893dc529e240d6a48