MediaRecorder: Upstream some moarrrr web_tests to WPT
This CL moves MediaRecorder-events-and-exceptions.html to the wpt/
folder, and refactors it to use canvas.captureStream() (ISO the
Chrome-specific blink_test_runner.cc thing).

It also fixes partially MediaRecorder-error.html: a new method
drawSomethingOnCanvas() is introduced and called to guarantee
that the canvas produces some captured VideoFrame, kicking the
event(s) generation. This fix is partial because Chrome doesn't
have MediaRecorderErrorEvent, but at least we verify that an
error event is thrown.

Bug: 919951
Change-Id: Ie9d7cfeeb2bf4452f4b5928eb2a7c644778751a7